feat(core): restructure package for improved code readability
Moves files around and renames directories to better organize the codebase and make it easier to navigate.

Breaking Changes:
None.
